Refactored delete and list commands.
